move database connection, util and response to a package
